client: fix bug in hijack
The goroutine that attaches to stdin should send an error in the channel
only if copy is called, otherwise hijack may return just because of a
nil error returned by the stdin goroutine. This doesn't affect the
behavior of all the background copying, but swallows errors in the
stdout goroutine.

Here's a build failure caused by this bug:
https://travis-ci.org/tsuru/tsuru/jobs/59943962#L3285.

Something changed at go tip and exposed this bug.