Issue http.ReadResponse hanging when running daemon with unix socket
On several occasions, the stream method hanged for specific container
(the scenario also repro in the docker CLI when running docker stats on
a specific container).
I tracked this error to the http.ReadResponse method but didn't find the
exact root cause in the docker deamon.
Using dial naitve timeout method should trigger a timeout (including a
unit test for the unix socket timeout scenario).