Update client.go
I didn't realize this was TLS-only until I tried to simplify my example. I had been looking for errors in Client.do(). After realizing it only happened in TLS, I looked at the TLS client code and found the problem. At the bottom of NewVersionedTLSClientFromBytes() you create a Transport with the TLS configuration. Transport uses keep-alive by default and will leave that connection open waiting for another request, but since each call to NewTLSClient() creates another new Transport, the connections are never reused.

The simplest fix is to disable keep-alive in the TLS transport. This has no performance impact over the current implementation because it wasn't using keep-alive anyway. A more performant solution would be to cache the TLS transport based on `endpoint` to be reused on subsequent calls.