determinism in hijack error returns
Previously, if both input and output streams were specified, it was random which
result was returned. So if the output write consistently failed, you could still
sometimes get a nil error back, because the input error chan value happened to
get selected first (and was nil).

Now, we explicitly wait for both errors, and return any non-nil error,
preferring the input error.