client: some tweaks in hijack
There's no need for the exit channel anymore, we're waiting for both
errors now (and they might be nil).

Closing the errChanOut is also no longer needed: the method always sends
a message on the channel.

Last, but not least, closing the input stream when the output stream
exists is important for releasing the other goroutine.