client: add method Endpoint
This method is useful for getting the endpoint when the user code
doesn't know it (for example, when the client is constructed from the
environment, using NewClientFromEnv).

I added a method instead of making the attribute exported to make it
clear that this method is intended for read-only access, exporting the
attribute could led to bad code, as users would be able to change the
value of Endpoint, but it would have no actual impact on the client,
as this attributes is not used in API calls.

Closes #380.