add support for searching private registries fixes #389
if found, send headers on GET requests

Change the SearchImages method to require authconfiguration. By passing
empty auth, it will still be able to search public images.

Searching works like this:
/image/search?term=nginx (Look for nginx in index.docker.io)
/image/search?term=quay.io/nginx (look for nginx in quay.io)

It is not possible to search private images on Docker Hub yet, but at
least this will ready the go-dockerclient when that is added!

Signed-off-by: Drew Wells <drew.wells00@gmail.com>