Use cleanhttp instead of http.Client
Unfortunately my previous commit overlooked that even in a fresh
http.Client{}, a shared DefaultTransport is used. We made a "cleanhttp"
library that makes it easy to get a clean client and/or transport that
still use the same default values as the Go standard library, so this
commit modifies my earlier changes to use that instead.