Fix stdin-only attach
Fix stdin-only attach so it doesn't wait for stdout/stderr to complete
before returning. If hijackOptions.stdout and hijackOptions.stderr are
nil, don't try to copy from the container's streams to ioutil.Discard,
as that keeps the attach session open as long as the container is
running. Instead, only copy stdout/stderr if they're non-nil.