client: use application/json content type in hijack
Docker 1.10.0 dropped support for plain/text on Exec. The client code
uses hijack both for exec and attach container. I've tested with all
versions of Docker from 1.7.1 to 1.10.0 and it worked as expected.

I should probably have an integration test for this one, will add in the
future. This should fix #455.