all: stop vendoring external dependencies
It's not good practice for a library, but we vendored our dependencies
in the past because go get github.com/fsouza/go-dockerclient was too
slow, but now vendoring is well advised in Go 1.5 and standard in Go
1.6. I'm also dropping Go 1.4 as some standard tools developed by the Go
core team already did that.

This will make tests on Travis run even faster, but I can hope that
golang/go#13078 will eventually get fixed.