Allow CreateNetworkOptions.IPAMOptions to be empty
If you do not do this and you just want default IPAM behavior, docker
daemon takes precisely 15 seconds to create you a network. It would seem
that some kind of discovery on a timeout is happening in the docker
daemon when you specify empty IPAM options. If you actually omit the
IPAM configuration (which is what docker/client does) then you get a
result immediately.