Support AuthConfiguration.IdentityToken
This change adds support for the sending and parsing of an
IdentityToken. IdentityToken's are retrieved via the /auth call (see
AuthCheck) and can then be used in place of the password. After a
successful `docker login` a users `.dockercfg` can contain an identity
token in place of the original username and password in the auth field.