Fix TestAuthConfigIdentityToken configuration
The docker config auths field is still expected to contain a base64
encoded string which is a colon ':' delimited username and password.
Even if the password is blank. This updates the test to include a
trailing colon in the auths payload. The password is empty as an
identitytoken should be provided.