Extend structs to have version tag, createImage as example
API changes can add additional parameters. To support these
correctly API versioned URLs are required.

PullImageOptions is extended to include a tag `ver:"1.32"`
for the Platform argument.

So PullImage will use the version tag of 1.32 if Platform is used
with a non-empty value, unless the client is configured to use a
later version, in which case the later version is used.

Using this feature required that the URL formation (getPath)
be constucted as a method on the client.

Internally streamUrl was created to taked the formed URL, rather
than the URL path (without versioning).

streamUrl calls checkAPIVersion purely to preserve API compatibility
that the client stucture of server version is populated on calling
PullImage/ImportImage.