Generate test TLS certificates on demand
This avoids having pem files that could expire and cause tests to
randomly start failing.

This makes tests slower, but computers are faster in 2025. If things are
really bad, we could refactor to use TestMain and make one certificate
rather than generating it on demand for tests, but I suspect this won't
be necessary.

Closes #1112.